fastapi/app/requirements.txt
fastapi
uvicorn
requests
boto3
pandas
python-dotenv
psycopg2-binary
sqlalchemy
mlflow
scikit-learn
numpy
